# src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "1.0.0",
    "title" : "",
    "contact" : {
      "name" : "",
      "url" : "",
      "email" : ""
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/hulkstore/api",
  "tags" : [ {
    "name" : "Producto",
    "description" : "Tag used to denote operations as private"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/productos" : {
      "get" : {
        "tags" : [ "Pproducto" ],
        "summary" : "Returns all the productos",
        "description" : "",
        "operationId" : "getProductos",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Productos found",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Producto"
              }
            }
          },
          "204" : {
            "description" : "No productos found"
          }
        }
      },
      "post" : {
        "tags" : [ "Producto" ],
        "summary" : "Creates a producto given a JSon Producto representation",
        "description" : "",
        "operationId" : "createProducto",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Producto to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Producto"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "The producto is created"
          },
          "415" : {
            "description" : "Format is not JSon"
          }
        }
      }
    },
    "/productos/count" : {
      "get" : {
        "tags" : [ "Producto" ],
        "summary" : "Returns the number of producto",
        "description" : "",
        "operationId" : "countProductos",
        "produces" : [ "text/plain" ],
        "responses" : {
          "200" : {
            "description" : "Number of productos found",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "204" : {
            "description" : "No productos found"
          }
        }
      }
    },
    "/productos/{id}" : {
      "get" : {
        "tags" : [ "Producto" ],
        "summary" : "Returns a producto given an id",
        "description" : "",
        "operationId" : "getProducto",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "minimum" : 1,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Producto found",
            "schema" : {
              "$ref" : "#/definitions/Producto"
            }
          },
          "400" : {
            "description" : "Invalid input. Id cannot be lower than 1"
          },
          "404" : {
            "description" : "Producto not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Producto" ],
        "summary" : "Deletes a Producto given an id",
        "description" : "",
        "operationId" : "deleteProducto",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "minimum" : 1,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Producto has been deleted"
          },
          "400" : {
            "description" : "Invalid input. Id cannot be lower than 1"
          },
          "500" : {
            "description" : "Producto not found"
          }
        }
      }
    }
  },
  "definitions" : {
    "Producto" : {
      "type" : "object",
      "required" : [ "isbn", "title" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Identifier"
        },
        "title" : {
          "type" : "string",
          "descripcion" : "Title of the Producto",
          "minLength" : 1,
          "maxLength" : 200
        },
        "description" : {
          "type" : "string",
          "description" : "Summary describing the Producto",
          "minLength" : 1,
          "maxLength" : 10000
        },
        "unitCost" : {
          "type" : "number",
          "format" : "float",
          "description" : "Unit cost",
          "minimum" : 1
        },
        "isbn" : {
          "type" : "string",
          "description" : "ISBN number",
          "minLength" : 1,
          "maxLength" : 50
        },
        "publicationDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date in which the Producto has been published"
        },
        "nbOfPages" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Number of pages"
        },
        "imageURL" : {
          "type" : "string",
          "description" : "URL of the image cover"
        },
        "language" : {
          "type" : "string",
          "description" : "",
          "enum" : [ "PLASTICO", "ACERO", "METAL", "ALGODON", "PORCELANA" ]
        }
      },
      "description" : "Producto resource representation"
    }
  }
}